[NFC] Restructure the project and move framework-related code into the `tyrell` package. This change makes packaging the Tyrell library a lot easier.
For external client of the framework, it now need to open up an
additional level of package when doing the imports. E.g. instead of
saying `import spec as S`, now we need to say `import tyrell.spec as S`.

For internal framework code, if it need to import other modules in the
framework, it need to use relative path rather than a global path. E.g.
in the `dsl/` module, instead of saying `import spec`, now we need to
say `import ..spec`.